Added the signin client code for ChatApplication
the signin is allowed through the email id and
user is asked for private key which is used to send
signed text "SecureChatServer" to the server and server
authenticates on basis of signed text if sign is valid
then authentication is successfull and token is given in response
else appropriate error is returned.
additional package is added to facilitate private-key download and
text signing functionalities.