Reworked whole chat system.
The messages are now encrypted on client using recipient's
public_key and sent messages are stored in localStorage
because we cannot retrive plain text after sending message.

the message is stored with message_id as key in localStorage
if message is recieved then we need to decrypt using our
private_key else if message was sent by us then it will already
be in our localStorage.